requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
# tensorflow>=2.16.0  # Optional: for autoencoder model

# Report generation
reportlab==4.0.7
openpyxl==3.1.2
xlsxwriter==3.1.9

# Utilities
python-dateutil==2.8.2
pydantic==2.5.2

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.1.3

# Database
supabase>=2.3.0
python-dotenv==1.0.0
websockets>=15.0
